+++ src/stars.rs
@@ -99,6 +99,18 @@ impl<'map> AnyStars<'map> {
         }
     }
 
+    /// Specify whether the map is a convert.
+    ///
+    /// This only needs to be specified if the map was converted manually beforehand.
+    #[inline]
+    pub fn is_convert(self, is_convert: bool) -> Self {
+        match self {
+            Self::Osu(_) | Self::Catch(_) => self,
+            Self::Taiko(t) => Self::Taiko(t.is_convert(is_convert)),
+            Self::Mania(m) => Self::Mania(m.is_convert(is_convert)),
+        }
+    }
+
     /// Consume the difficulty calculator and calculate
     /// difficulty attributes for the given parameters.
     #[inline]